CSS: Add fallback to `font-family` property
In the CSS stylesheet `temple-of-rust.css`, in the rule for the `body`
element, the property `font-family` was specified with only a single
value, `'Fira Mono'`, with no fallback value for use in the event that
the typeface Fira Mono is unavailable; for me, this resulted in the
text being rendered in the typeface Liberation Serif.

This commit changes the afore-described `font-family` property to fall
back to the viewer's default monospace typeface if Fira Mono is
unavailable.